requirements.txt
django~=5.1.2
psycopg2-binary

django-debug-toolbar~=4.4.6